todx-dl/requirements.txt
--find-links https://download.pytorch.org/whl/cu118
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1
torchinfo==1.8.0
lightning==2.0.7
torchmetrics==1.1.0
tqdm==4.66.1

# common requirements
kafka-python==2.0.2
python-box[all]==7.1.1